added build.xml and build.properties to vesper-history
