+++ build.properties
@@ -2,6 +2,7 @@ lib.dir=lib
 src.dir=src
 test.dir=test
 cli.src.dir=ext/vesper-cli/src
+history.src.dir=ext/vesper-history/src
 build.dir=build
 javadoc.packagenames=edu.ucsc.refactor,edu.ucsc.refactor.spi,\
   edu.ucsc.refactor.util,\
