+++ ext/vesper-history/build.properties
@@ -1,0 +1,6 @@
+lib.dir=../../lib
+src.dir=src
+test.dir=test
+build.dir=build
+test.class=edu.ucsc.refactor.AllHistoryTests
+module=edu.ucsc.refactor
