+++ ext/vesper-history/lib/org.eclipse.egit.github.core-3.3.0-SNAPSHOT.jar
