+++ tasks.xml
@@ -92,7 +92,6 @@
             <zipfileset excludes="META-INF/**/*" src="${common.basedir}/lib/org.eclipse.core.jobs_3.5.300.v20130429-1813.jar"/>
             <zipfileset excludes="META-INF/**/*" src="${common.basedir}/lib/org.eclipse.core.resources_3.8.100.v20130521-2026.jar"/>
             <zipfileset excludes="META-INF/**/*" src="${common.basedir}/lib/org.eclipse.core.runtime_3.9.0.v20130326-1255.jar"/>
-            <zipfileset excludes="META-INF/**/*" src="${common.basedir}/lib/org.eclipse.egit.github.core-3.3.0-SNAPSHOT-sources.jar"/>
             <zipfileset excludes="META-INF/**/*" src="${common.basedir}/lib/org.eclipse.equinox.common_3.6.200.v20130402-1505.jar"/>
             <zipfileset excludes="META-INF/**/*" src="${common.basedir}/lib/org.eclipse.equinox.preferences_3.5.100.v20130422-1538.jar"/>
             <zipfileset excludes="META-INF/**/*" src="${common.basedir}/lib/org.eclipse.jdt.core_3.9.0.v20130604-1421.jar"/>
